+++ lib/xproject_generator.dart
@@ -38,61 +38,79 @@ class ProjectGenerator {
     logger.info('📝 Generating files...');
     await FileGenerator.generate(config);
 
-    logger
-        .success('\n✅ Project created successfully at: ${config.projectPath}');
+    logger.success(
+      '\n✅ Project created successfully at: ${config.projectPath}',
+    );
     logger.info('\n 📝 Next steps:');
     logger.info('  - cd ${config.appName}');
     logger.info('  - flutter pub get');
     logger.info(
-        '\n 🔑 Setup secret keys for SecureCompressor manually and generate secretkey.dart file using secret_key_scrypt_generator package');
+      '\n 🔑 Setup secret keys for SecureCompressor manually and generate secretkey.dart file using secret_key_scrypt_generator package',
+    );
     logger.info('  - Add your secret keys in secretkey.dart file located at:');
     logger.info('    lib/core/config/secretkey.dart');
     logger.info(
-        '  - Refer to https://pub.dev/packages/secret_key_scrypt_generator for generating secretkey.dart file');
+      '  - Refer to https://pub.dev/packages/secret_key_scrypt_generator for generating secretkey.dart file',
+    );
     if (config.useFirebase) {
       logger.info('\n 🔥 Firebase setup:');
       logger.info(
-          '  - Place your google-services.json file at: android/app/google-services.json');
+        '  - Place your google-services.json file at: android/app/google-services.json',
+      );
       logger.info(
-          '  - Place your GoogleService-Info.plist file at: ios/Runner/GoogleService-Info.plist');
+        '  - Place your GoogleService-Info.plist file at: ios/Runner/GoogleService-Info.plist',
+      );
       logger.info(
-          '  - or use flutterfire CLI to configure Firebase automatically');
+        '  - or use flutterfire CLI to configure Firebase automatically',
+      );
       logger.info(
-          '  - For more details, refer to the Firebase setup guide in the README.md file');
+        '  - For more details, refer to the Firebase setup guide in the README.md file',
+      );
     }
     logger.info('\n 🔄 Sync freezed and other Build Runners');
     logger.info('   - Make sure the code was clear (No errors)');
     logger.info(
-        '   - Run task with Cmd/Ctrl + Shift + P → Tasks: Run Task -> Synchronize Build Runners');
-    logger.info('\n 📱 iOS one-time setup (required for --flavor):');
-    logger.info('\n 📱 Enhance Runner configuration for multiple flavors:');
+      '   - Run task with Cmd/Ctrl + Shift + P → Tasks: Run Task -> Synchronize Build Runners',
+    );
+    logger.info('\n 📱 iOS one-time setup (required for --flavor)');
+    logger.info('\n Enhance Runner configuration for multiple flavors:');
     logger.info('  1. Open ios/Runner.xcodeproj in Xcode');
     logger.info('  2. Select Runner on PROJECT → Info');
     logger.info(
-        '  3. Under "Configurations", duplicate "Debug", "Profile", and "Release" for each flavor: development, staging');
+      '  3. Under "Configurations", duplicate "Debug", "Profile", and "Release" for each flavor: development, staging',
+    );
     logger.info('     (e.g., Debug-development, Release-development, etc.)');
     logger.info(
-        '  4. Rename origin configurations intu "Debug-production", "Profile-production", and "Release-production"');
+      '  4. Rename origin configurations into "Debug-production", "Profile-production", and "Release-production"',
+    );
     logger.info('  5. Go to "Build Settings" in TARGETS → Runner');
     logger.info('  6. Search for "Product Bundle Identifier"');
     logger.info(
-        '  7. Set the bundle identifier for each configuration accordingly');
+      '  7. Set the bundle identifier for each configuration accordingly',
+    );
     logger.info(
-        '     (e.g., com.example.app.development, com.example.app.staging, etc.)');
+      '     (e.g., com.example.app.development, com.example.app.staging, etc.)',
+    );
     logger.info(
-        '  8. Setup display name for each configuration environment if needed.');
+      '  8. Setup display name for each configuration environment if needed.',
+    );
     logger.info('     - Tap Add icon to add new User-Defined Setting');
     logger.info(
-        '     - Set "APP_DISPLAY_NAME" as the key and your desired app name as the value for each configuration');
+      '     - Set "APP_DISPLAY_NAME" as the key and your desired app name as the value for each configuration',
+    );
     logger.info(
-        '     - e.g., "Test App Dev" for development, "Test App Staging" for staging, etc.');
+      '     - e.g., "Test App Dev" for development, "Test App Staging" for staging, etc.',
+    );
     logger.info(
-        '     - Open Info.plist and set the Bundle display name to \$(APP_DISPLAY_NAME)');
+      '     - Open Info.plist and set the Bundle display name to \$(APP_DISPLAY_NAME)',
+    );
     logger.info('');
     logger.info(
-        ' 📃 This project uses an iOS Firebase script located at: ios/firebase.sh');
+      ' 📃 This project uses an iOS Firebase script located at: ios/firebase.sh',
+    );
     logger.info(
-        '  Please complete the following steps in Xcode to complete the iOS flavoring setup:');
+      '  Please complete the following steps in Xcode to complete the iOS flavoring setup:',
+    );
     logger.info('  1. Open Xcode workspace:');
     logger.info('     open ios/Runner.xcworkspace');
     logger.info('  2. Select Runner → Build Phases');
@@ -100,14 +118,16 @@ class ProjectGenerator {
     logger.info('  4. Paste this script path:');
     logger.info('     "ios/firebase.sh"');
     logger.info('  5. Rename script to "Firebase Script"');
-    logger
-        .info('  6. Make sure this phase runs BEFORE "Copy Bundle Resources"');
+    logger.info(
+      '  6. Make sure this phase runs BEFORE "Copy Bundle Resources"',
+    );
     logger.info('  7. Create schemes: development, staging, production');
     logger.info('     (Product → Scheme → Manage Schemes)');
 
     logger.info('\n ▶ Run app:');
     logger.info(
-        '  flutter run --flavor development -t lib/main_development.dart');
+      '  flutter run --flavor development -t lib/main_development.dart',
+    );
   }
 
   /// Collect user input for project creation (name, package, state management, etc.).
@@ -148,9 +168,7 @@ class ProjectGenerator {
 
         // 1️⃣ IZINKAN INPUT SAAT MENGETIK
         if (!RegExp(r'^[a-z0-9._]+$').hasMatch(x)) {
-          print(
-            '  ⚠️  Use lowercase letters, numbers, underscores, and dots',
-          );
+          print('  ⚠️  Use lowercase letters, numbers, underscores, and dots');
           return false;
         }
 
@@ -171,9 +189,7 @@ class ProjectGenerator {
 
         // allow typing
         if (!RegExp(r'^[a-z0-9.-]+$').hasMatch(x)) {
-          print(
-            '  ⚠️  Use lowercase letters, numbers, dots, and hyphens',
-          );
+          print('  ⚠️  Use lowercase letters, numbers, dots, and hyphens');
           return false;
         }
 
@@ -188,13 +204,13 @@ class ProjectGenerator {
 
     final iosBundleId = iosInput.isEmpty ? androidPackage : iosInput;
 
-    logger.success(
-      'iOS bundle ID · $iosBundleId',
-    );
+    logger.success('iOS bundle ID · $iosBundleId');
 
-    final firebaseInput =
-        Input(prompt: 'Use Firebase? (Y/n): ').interact().trim();
-    final useFirebase = firebaseInput.isEmpty ||
+    final firebaseInput = Input(
+      prompt: 'Use Firebase? (Y/n): ',
+    ).interact().trim();
+    final useFirebase =
+        firebaseInput.isEmpty ||
         firebaseInput.toLowerCase() == 'y' ||
         firebaseInput.toLowerCase() == 'yes';