+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.1
+
+- Add customable TextStyle function using Inter
+- Add customable AppColor
+- Add api request boilerplate
+- Bug Fixing
+
 ## 1.0.0
 
 - Initial version.